interface with stockfish for eval
I realized that the real challenge is understanding the current position
of the board, and once I have a representation, it would make sense to
use an efficient chess engine to analyze what moves would normally be
good. This may need to be messed around with later, but for now I am
using Stockfish.